@@ -1,125 +1,96 @@
 name: smooth_app
 description: A new official Open Food Facts application build with great user experience in mind.
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-
-#{major}.{minor}.{patch}+{version}
-# this version can be ignored as it is automatically increased during the release process
-version: 0.0.1+5
-
+version: 0.0.0
+publish_to: "none"
 
 environment:
-  sdk: ">=2.10.4 <3.0.0"
-
+  sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
-  carousel_slider: ^4.0.0-nullsafety.0
-  cupertino_icons: ^1.0.2
-  device_preview: 0.7.1
-  diacritic: ^0.1.3
+  flutter_lints: ^1.0.4
+  carousel_slider: ^4.0.0
+  cupertino_icons: ^1.0.4
+  device_preview: ^1.0.0
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_staggered_animations: ^1.0.0
-  flutter_staggered_grid_view: ^0.4.0
-  flutter_svg: ^0.21.0+1
-  flutter_typeahead: ^3.1.1
-  image_cropper: ^1.4.0
-  image_picker: ^0.7.4
-  modal_bottom_sheet: ^2.0.0
-  openfoodfacts: ^1.1.0-beta
-  # Uncomment those lines if you want to use a local version of the openfoodfacts package
-  #openfoodfacts:
-  #  path: ../../../openfoodfacts-dart
-  package_info: ^2.0.0
-  photo_view: ^0.11.1
-  provider: ^5.0.0
-  qr_code_scanner: ^0.4.0
-  rubber: ^1.0.0
-  sentry: ^5.0.0
-  share: ^2.0.1
-  smooth_ui_library:
-    path: ../smooth_ui_library
-  sqflite: ^2.0.0+3
-  url_launcher: ^6.0.3
+  flutter_svg: ^1.0.3
+  flutter_map: ^0.14.0
+  flutter_widget_from_html_core: ^0.8.5+1
+  flutter_secure_storage: ^5.0.2
+  hive: ^2.0.6
+  hive_flutter: ^1.1.0
+  http: ^0.13.4
+  image_picker: ^0.8.4+10
+  iso_countries: 2.1.0
+  latlong2: ^0.8.1
+  matomo_forever: ^1.0.0+1
+  modal_bottom_sheet: ^2.0.1
+  openfoodfacts: ^1.13.0
+#  openfoodfacts:
+#    path: ../../../openfoodfacts-dart
+  package_info_plus: ^1.4.0
+  permission_handler: ^9.2.0
+  photo_view: ^0.13.0
+  uuid: ^3.0.6
+  provider: ^6.0.2
+  qr_code_scanner: ^0.7.0
+  rubber: ^1.0.1
+  sentry_flutter: ^6.3.0
+  url_launcher: ^6.0.20
+  visibility_detector: ^0.2.2
+  google_ml_barcode_scanner: ^0.0.2
+  camera: ^0.9.4+16
+  percent_indicator: ^4.0.0
+  crop_your_image: ^0.7.0
+  mailto: ^2.0.0
+  flutter_native_splash: ^2.1.0
 
+
+  
 dev_dependencies:
-  flutter_native_splash: ^1.1.7+1
+  flutter_launcher_icons: ^0.9.2
   flutter_test:
     sdk: flutter
+  mockito: ^5.1.0
+  path: ^1.8.0 # careful with 1.8.1 because of https://github.com/flutter/flutter/issues/95478
+
+# 'flutter pub run flutter_launcher_icons:main' to update
+flutter_icons:
+  android: "launcher_icon"
+  ios: true
+  remove_alpha_ios: true
+  image_path: "assets/app/smoothie-icon.1200x1200.png"
+  adaptive_icon_background: "#00ccff"
+  # Only the inner 72x72dp of the 108x108dp adaptive icon is shown
+  # (extra padding of 18dp on all sides is used for visual effects)
+  # https://developer.android.com/guide/practices/ui_guidelines/icon_design_adaptive
+  adaptive_icon_foreground: "assets/app/smoothie-icon-transparent.android-adaptive.1080x1080.png"
 
 # 'flutter pub run flutter_native_splash:create' to update
 flutter_native_splash:
-  image: assets/temp/Logo.png
-  color: "ffffff"
-  image_dark: assets/temp/Logo.png
-  color_dark: "#181818"
+  image: assets/app/smoothie-icon-transparent.1200x1200.png
+  color: "00ccff"
+  image_dark: assets/app/smoothie-icon-transparent.1200x1200.png
+  color_dark: "004455"
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
   generate: true
 
   assets:
     - assets/actions/
+    - assets/app/
     - assets/cache/
     - assets/categories/
     - assets/data/
+    - assets/home/
     - assets/labels/
     - assets/metadata/
     - assets/misc/
     - assets/navigation/
+    - assets/onboarding/
     - assets/product/
     - assets/ikonate_thin/
     - assets/ikonate_bold/
-    - assets/temp/
-
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
-
-publish_to: 'none'
