@@ -1,88 +1,55 @@
-// Dart imports:
 import 'dart:io';
 
-// Flutter imports:
 import 'package:flutter/material.dart';
-
-// Package imports:
-import 'package:image_cropper/image_cropper.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:openfoodfacts/model/Product.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
-import 'package:openfoodfacts/utils/LanguageHelper.dart';
-
-import 'package:smooth_app/pages/product/product_image_page.dart';
+import 'package:smooth_app/data_models/product_image_data.dart';
+import 'package:smooth_app/helpers/picture_capture_helper.dart';
+import 'package:smooth_app/pages/image_crop_page.dart';
+import 'package:smooth_app/pages/product/product_image_gallery_view.dart';
 
 class ImageUploadCard extends StatefulWidget {
   const ImageUploadCard({
-    this.product,
-    this.imageField,
-    this.imageUrl,
-    this.title,
-    this.buttonText,
+    required this.product,
+    required this.productImageData,
+    required this.allProductImagesData,
+    required this.onUpload,
   });
 
   final Product product;
-  final ImageField imageField;
-  final String imageUrl;
-  final String title;
-  final String buttonText;
+  final ProductImageData productImageData;
+  final List<ProductImageData> allProductImagesData;
+  final Function(BuildContext) onUpload;
 
   @override
-  _ImageUploadCardState createState() => _ImageUploadCardState();
+  State<ImageUploadCard> createState() => _ImageUploadCardState();
 }
 
 class _ImageUploadCardState extends State<ImageUploadCard> {
-  ImageProvider _imageProvider; // Normal size image to display in carousel
-  ImageProvider
+  ImageProvider? _imageProvider; // Normal size image to display in carousel
+  ImageProvider?
       _imageFullProvider; // Full resolution image to display in image page
 
   Future<void> _getImage() async {
-    final ImagePicker picker = ImagePicker();
+    final File? croppedImageFile = await startImageCropping(context);
 
-    final PickedFile pickedFile =
-        await picker.getImage(source: ImageSource.camera);
+    if (croppedImageFile != null) {
+      setState(() {
+        // Update the image to load the new image file
+        // The same full resolution image is used for both the carousel and the image page
+        _imageProvider = FileImage(croppedImageFile);
+        _imageFullProvider = _imageProvider;
+      });
 
-    if (pickedFile != null) {
-      final File croppedImageFile = await ImageCropper.cropImage(
-        sourcePath: pickedFile.path,
-        androidUiSettings: const AndroidUiSettings(
-          lockAspectRatio: false,
-          hideBottomControls: true,
-        ),
+      final bool isUploaded = await uploadCapturedPicture(
+        context,
+        barcode: widget.product
+            .barcode!, //Probably throws an error, but this is not a big problem when we got a product without a barcode
+        imageField: widget.productImageData.imageField,
+        imageUri: croppedImageFile.uri,
       );
-
-      if (croppedImageFile != null) {
-        setState(() {
-          // Update the image to load the new image file
-          // The same full resolution image is used for both the carousel and the image page
-          _imageProvider = FileImage(croppedImageFile);
-          _imageFullProvider = _imageProvider;
-        });
-
-        final SendImage image = SendImage(
-          lang: LanguageHelper.fromJson(
-              Localizations.localeOf(context).languageCode),
-          barcode: widget.product.barcode,
-          imageField: widget.imageField,
-          imageUri: croppedImageFile.uri,
-        );
-
-        // a registered user login for https://world.openfoodfacts.org/ is required
-        //ToDo: Add user
-        const User myUser =
-            User(userId: 'smoothie-app', password: 'strawberrybanana');
-
-        // query the OpenFoodFacts API
-        final Status result =
-            await OpenFoodAPIClient.addProductImage(myUser, image);
-
-        if (result.status != 'status ok') {
-          throw Exception('image could not be uploaded: ' +
-              result.error +
-              ' ' +
-              result.imageId.toString());
-        }
+      croppedImageFile.delete();
+      if (isUploaded) {
+        widget.onUpload(context);
       }
     }
   }
@@ -92,35 +59,34 @@ class _ImageUploadCardState extends State<ImageUploadCard> {
     // We can already have an _imageProvider for a file that is going to be uploaded
     // or an imageUrl for a network image
     // or no image yet
-
-    if ((_imageProvider == null) && (widget.imageUrl != null)) {
-      _imageProvider = NetworkImage(widget.imageUrl);
+    if ((_imageProvider == null) &&
+        (widget.productImageData.imageUrl != null)) {
+      _imageProvider = NetworkImage(widget.productImageData.imageUrl!);
     }
 
     if (_imageProvider != null) {
       return GestureDetector(
         child: Center(
             child:
-                Image(image: _imageProvider, fit: BoxFit.cover, height: 1000)),
+                Image(image: _imageProvider!, fit: BoxFit.cover, height: 1000)),
         onTap: () {
           // if _imageFullProvider is null, we are displaying a small network image in the carousel
           // we need to load the full resolution image
 
           if (_imageFullProvider == null) {
             final String _imageFullUrl =
-                widget.imageUrl.replaceAll('.400.', '.full.');
+                widget.productImageData.imageUrl!.replaceAll('.400.', '.full.');
             _imageFullProvider = NetworkImage(_imageFullUrl);
           }
 
           Navigator.push<Widget>(
             context,
             MaterialPageRoute<Widget>(
-              builder: (BuildContext context) => ProductImagePage(
-                  product: widget.product,
-                  imageField: widget.imageField,
-                  imageProvider: _imageFullProvider,
-                  title: widget.title,
-                  buttonText: widget.buttonText),
+              builder: (BuildContext context) => ProductImageGalleryView(
+                productImageData: widget.productImageData,
+                allProductImagesData: widget.allProductImagesData,
+                title: widget.productImageData.title,
+              ),
             ),
           );
         },
@@ -129,7 +95,7 @@ class _ImageUploadCardState extends State<ImageUploadCard> {
       return ElevatedButton.icon(
         onPressed: _getImage,
         icon: const Icon(Icons.add_a_photo),
-        label: Text(widget.buttonText),
+        label: Text(widget.productImageData.buttonText),
       );
     }
   }
