@@ -1,11 +1,8 @@
-// Dart imports:
-import 'dart:ui';
-
-// Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:smooth_app/generic_lib/design_constants.dart';
 
 class SmoothProductCardLoading extends StatelessWidget {
-  const SmoothProductCardLoading({@required this.barcode});
+  const SmoothProductCardLoading({required this.barcode});
 
   final String barcode;
 
@@ -14,7 +11,7 @@ class SmoothProductCardLoading extends StatelessWidget {
     return Container(
       decoration: BoxDecoration(
         color: Theme.of(context).colorScheme.surface,
-        borderRadius: const BorderRadius.all(Radius.circular(15.0)),
+        borderRadius: ROUNDED_BORDER_RADIUS,
       ),
       child: Column(
         mainAxisSize: MainAxisSize.max,
