@@ -1,219 +1,139 @@
-// Flutter imports:
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-// Package imports:
 import 'package:openfoodfacts/model/Product.dart';
+import 'package:openfoodfacts/utils/PnnsGroups.dart';
 
-// Project imports:
-import 'package:smooth_app/themes/smooth_theme.dart';
+enum ProductListType {
+  /// API search for [PnnsGroup2Filter] related food groups
+  HTTP_SEARCH_GROUP,
+
+  /// API search by [SearchTerms] keywords
+  HTTP_SEARCH_KEYWORDS,
+
+  /// API search for [CategoryProductQuery] category
+  HTTP_SEARCH_CATEGORY,
+
+  /// Current scan session; can be easily cleared by the end-user
+  SCAN_SESSION,
+
+  /// History of products seen by the end-user
+  HISTORY,
+}
+
+extension ProductListTypeExtension on ProductListType {
+  static const Map<ProductListType, String> _keys = <ProductListType, String>{
+    ProductListType.HTTP_SEARCH_GROUP: 'http/search/group',
+    ProductListType.HTTP_SEARCH_KEYWORDS: 'http/search/keywords',
+    ProductListType.HTTP_SEARCH_CATEGORY: 'http/search/category',
+    ProductListType.SCAN_SESSION: 'scan_session',
+    ProductListType.HISTORY: 'history',
+  };
+
+  String get key => _keys[this]!;
+}
 
 class ProductList {
-  ProductList({
-    @required this.listType,
-    @required this.parameters,
-    this.databaseTimestamp,
-    this.databaseCount,
-    this.databaseCountDistinct,
-  });
+  ProductList._({required this.listType, this.parameters = ''});
 
-  static const String _EXTRA_COLOR = 'color';
-  static const String _EXTRA_ICON = 'icon';
+  ProductList.keywordSearch(final String keywords)
+      : this._(
+          listType: ProductListType.HTTP_SEARCH_KEYWORDS,
+          parameters: keywords,
+        );
 
-  static const String _COLOR_RED = 'red';
-  static const String _COLOR_ORANGE = 'orange';
-  static const String _COLOR_GREEN = 'green';
-  static const String _COLOR_BLUE = 'blue';
-  static const String _COLOR_PURPLE = 'purple';
+  ProductList.categorySearch(final String category)
+      : this._(
+          listType: ProductListType.HTTP_SEARCH_CATEGORY,
+          parameters: category,
+        );
 
-  static const Map<String, MaterialColor> _COLORS = <String, MaterialColor>{
-    _COLOR_RED: Colors.red,
-    _COLOR_ORANGE: Colors.orange,
-    _COLOR_GREEN: Colors.green,
-    _COLOR_BLUE: Colors.blue,
-    _COLOR_PURPLE: Colors.purple,
-  };
+  ProductList.groupSearch(final PnnsGroup2 group)
+      : this._(
+          listType: ProductListType.HTTP_SEARCH_GROUP,
+          parameters: group.id,
+        );
 
-  static const List<String> ORDERED_COLORS = <String>[
-    _COLOR_RED,
-    _COLOR_ORANGE,
-    _COLOR_GREEN,
-    _COLOR_BLUE,
-    _COLOR_PURPLE,
-  ];
+  ProductList.history() : this._(listType: ProductListType.HISTORY);
 
-  static const Map<String, String> _DEFAULT_COLOR_TAG_PER_TYPE =
-      <String, String>{
-    LIST_TYPE_HTTP_SEARCH_KEYWORDS: _COLOR_RED,
-    LIST_TYPE_HTTP_SEARCH_CATEGORY: _COLOR_BLUE,
-    LIST_TYPE_HTTP_SEARCH_GROUP: _COLOR_ORANGE,
-    LIST_TYPE_SCAN: _COLOR_GREEN,
-    LIST_TYPE_HISTORY: _COLOR_BLUE,
-    LIST_TYPE_USER_DEFINED: _COLOR_PURPLE,
-  };
+  ProductList.scanSession() : this._(listType: ProductListType.SCAN_SESSION);
 
-  static const String _ICON_TAG = 'tag';
-  static const String _ICON_HEART = 'heart';
-  static const String _ICON_SEARCH = 'search';
-  static const String _ICON_CATEGORY = 'category';
-  static const String _ICON_GROUP = 'group';
-  static const String _ICON_SCAN = 'scan';
-  static const String _ICON_HISTORY = 'history';
-
-  static const Map<String, List<String>> _ORDERED_ICONS_PER_TYPE =
-      <String, List<String>>{
-    LIST_TYPE_HTTP_SEARCH_KEYWORDS: <String>[_ICON_SEARCH],
-    LIST_TYPE_HTTP_SEARCH_CATEGORY: <String>[_ICON_CATEGORY],
-    LIST_TYPE_HTTP_SEARCH_GROUP: <String>[_ICON_GROUP],
-    LIST_TYPE_SCAN: <String>[_ICON_SCAN],
-    LIST_TYPE_HISTORY: <String>[_ICON_HISTORY],
-    LIST_TYPE_USER_DEFINED: <String>[_ICON_TAG, _ICON_HEART]
-  };
-
-  static const Map<String, IconData> _ICON_DATA = <String, IconData>{
-    _ICON_TAG: CupertinoIcons.tag_fill,
-    _ICON_HEART: CupertinoIcons.heart_fill,
-    _ICON_SEARCH: Icons.search,
-    _ICON_CATEGORY: Icons.description,
-    _ICON_GROUP: Icons.fastfood,
-    _ICON_SCAN: CupertinoIcons.barcode,
-    _ICON_HISTORY: Icons.history,
-  };
-
-  static const Map<String, String> _DEFAULT_ICON_TAG_PER_TYPE =
-      <String, String>{
-    LIST_TYPE_HTTP_SEARCH_KEYWORDS: _ICON_SEARCH,
-    LIST_TYPE_HTTP_SEARCH_CATEGORY: _ICON_CATEGORY,
-    LIST_TYPE_HTTP_SEARCH_GROUP: _ICON_GROUP,
-    LIST_TYPE_SCAN: _ICON_SCAN,
-    LIST_TYPE_HISTORY: _ICON_HISTORY,
-    LIST_TYPE_USER_DEFINED: _ICON_TAG,
-  };
-
-  final String listType;
+  final ProductListType listType;
   final String parameters;
-  final int databaseTimestamp;
-  final int databaseCount;
-  final int databaseCountDistinct;
-  Map<String, String> extraTags;
 
   final List<String> _barcodes = <String>[];
   final Map<String, Product> _products = <String, Product>{};
 
+  /// API search for [PnnsGroup2Filter] related food groups
   static const String LIST_TYPE_HTTP_SEARCH_GROUP = 'http/search/group';
+
+  /// API search by [SearchTerms] keywords
   static const String LIST_TYPE_HTTP_SEARCH_KEYWORDS = 'http/search/keywords';
+
+  /// API search for [CategoryProductQuery] category
   static const String LIST_TYPE_HTTP_SEARCH_CATEGORY = 'http/search/category';
-  static const String LIST_TYPE_SCAN = 'scan';
+
+  /// Current scan session; can be easily cleared by the end-user
+  static const String LIST_TYPE_SCAN_SESSION = 'scan_session';
+
+  /// History of products seen by the end-user
   static const String LIST_TYPE_HISTORY = 'history';
-  static const String LIST_TYPE_USER_DEFINED = 'user';
 
   List<String> get barcodes => _barcodes;
 
-  set colorTag(final String tag) => _setExtra(_EXTRA_COLOR, tag);
-  set iconTag(final String tag) => _setExtra(_EXTRA_ICON, tag);
-
-  void _setExtra(final String key, final String value) {
-    extraTags ??= <String, String>{};
-    extraTags[key] = value;
-  }
-
-  String get _colorTag =>
-      (extraTags == null ? null : extraTags[_EXTRA_COLOR]) ??
-      _DEFAULT_COLOR_TAG_PER_TYPE[listType];
-  String get _iconTag =>
-      (extraTags == null ? null : extraTags[_EXTRA_ICON]) ??
-      _DEFAULT_ICON_TAG_PER_TYPE[listType];
-
-  List<String> getPossibleIcons() => _ORDERED_ICONS_PER_TYPE[listType];
-
   bool isEmpty() => _barcodes.isEmpty;
 
-  void clear() {
-    _barcodes.clear();
-    _products.clear();
-  }
+  Product getProduct(final String barcode) => _products[barcode]!;
 
-  Product getProduct(final String barcode) => _products[barcode];
-
-  String get lousyKey =>
-      '$listType/$parameters'; // TODO(monsieurtanuki): does not work if you change the name
-
-  IconData get iconData => _ICON_DATA[_iconTag] ?? _ICON_DATA[_ICON_TAG];
-
-  static Widget getReferenceIcon({
-    @required final ColorScheme colorScheme,
-    @required final String colorTag,
-    @required final String iconTag,
-    @required final ColorDestination colorDestination,
-  }) =>
-      getTintedIcon(
-        colorScheme: colorScheme,
-        materialColor: _getReferenceMaterialColor(colorTag),
-        iconData: _ICON_DATA[iconTag] ?? _ICON_DATA[_ICON_TAG],
-        colorDestination: colorDestination,
-      );
-
-  static Widget getTintedIcon({
-    @required final ColorScheme colorScheme,
-    @required final MaterialColor materialColor,
-    @required final IconData iconData,
-    @required final ColorDestination colorDestination,
-  }) =>
-      Icon(
-        iconData,
-        color: colorDestination == null
-            ? null
-            : SmoothTheme.getColor(
-                colorScheme, materialColor, colorDestination),
-      );
-
-  static MaterialColor _getReferenceMaterialColor(final String colorTag) =>
-      _COLORS[colorTag] ?? _COLORS[_COLOR_RED];
-
-  Widget getIcon(
-    final ColorScheme colorScheme,
-    final ColorDestination colorDestination,
-  ) =>
-      getReferenceIcon(
-        colorScheme: colorScheme,
-        colorTag: _colorTag,
-        iconTag: _iconTag,
-        colorDestination: colorDestination,
-      );
-
-  MaterialColor getMaterialColor() => _getReferenceMaterialColor(_colorTag);
+  bool isSameAs(final ProductList other) =>
+      listType == other.listType && parameters == other.parameters;
 
   void refresh(final Product product) {
-    if (product == null) {
-      throw Exception('null product');
-    }
-    final String barcode = product.barcode;
+    final String? barcode = product.barcode;
     if (barcode == null) {
       throw Exception('null barcode');
     }
     _products[barcode] = product;
   }
 
-  void add(final Product product) {
-    refresh(product);
-    _barcodes.add(product.barcode);
+  /// Removes a barcode from the list
+  ///
+  /// Returns false if not already in the list
+  /// Don't forget to update the database afterwards
+  bool remove(final String barcode) {
+    if (!_barcodes.contains(barcode)) {
+      return false;
+    }
+    _barcodes.remove(barcode);
+    _products.remove(barcode);
+    return true;
   }
 
-  void addAll(final List<Product> products) => products.forEach(add);
+  /// Sets all products with the same order as the input list
+  void setAll(final List<Product> products) {
+    final List<String> barcodes = <String>[];
+    final Map<String, Product> productMap = <String, Product>{};
+    for (final Product product in products) {
+      final String barcode = product.barcode!;
+      barcodes.add(barcode);
+      productMap[barcode] = product;
+    }
+    set(barcodes, productMap);
+  }
 
   void set(
     final List<String> barcodes,
     final Map<String, Product> products,
   ) {
-    clear();
-    _barcodes.addAll(barcodes);
+    _barcodes.clear();
+    _products.clear();
     _products.addAll(products);
+    _barcodes.addAll(barcodes);
   }
 
   List<Product> getList() {
     final List<Product> result = <Product>[];
-    for (final String barcode in _barcodes) {
-      final Product product = _products[barcode];
+    final Iterable<String> barcodes =
+        _isReversed() ? _barcodes.reversed : _barcodes;
+    for (final String barcode in barcodes) {
+      final Product? product = _products[barcode];
       if (product == null) {
         throw Exception('no product for barcode $barcode');
       }
@@ -222,33 +142,15 @@ class ProductList {
     return result;
   }
 
-  List<Product> getUniqueList() {
-    final List<Product> result = <Product>[];
-    final List<String> orderedBarcodes = getOrderedBarcodes();
-    for (final String barcode in orderedBarcodes) {
-      final Product product = _products[barcode];
-      if (product == null) {
-        throw Exception('no product for barcode $barcode');
-      }
-      result.add(product);
+  bool _isReversed() {
+    switch (listType) {
+      case ProductListType.HTTP_SEARCH_GROUP:
+      case ProductListType.HTTP_SEARCH_KEYWORDS:
+      case ProductListType.HTTP_SEARCH_CATEGORY:
+        return false;
+      case ProductListType.SCAN_SESSION:
+      case ProductListType.HISTORY:
+        return true;
     }
-    return result;
   }
-
-  List<String> getOrderedBarcodes() {
-    final List<String> result = <String>[];
-    final Iterable<String> iterable =
-        _isReversed ? barcodes.reversed : barcodes;
-    for (final String barcode in iterable) {
-      if (result.contains(barcode)) {
-        continue;
-      }
-      result.add(barcode);
-    }
-    return result;
-  }
-
-  bool get _isReversed =>
-      listType == ProductList.LIST_TYPE_HISTORY ||
-      listType == ProductList.LIST_TYPE_SCAN;
 }
