@@ -1,29 +1,39 @@
-import 'package:http/http.dart' as http;
+import 'dart:async' show Future;
+
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+import 'package:openfoodfacts/model/Attribute.dart';
+import 'package:openfoodfacts/model/AttributeGroup.dart';
 import 'package:openfoodfacts/personalized_search/available_attribute_groups.dart';
 import 'package:openfoodfacts/personalized_search/available_preference_importances.dart';
 import 'package:openfoodfacts/personalized_search/available_product_preferences.dart';
 import 'package:openfoodfacts/personalized_search/preference_importance.dart';
 import 'package:openfoodfacts/personalized_search/product_preferences_manager.dart';
 import 'package:openfoodfacts/personalized_search/product_preferences_selection.dart';
-import 'package:openfoodfacts/model/Attribute.dart';
-import 'package:openfoodfacts/model/AttributeGroup.dart';
+import 'package:smooth_app/data_models/downloadable_string.dart';
+import 'package:smooth_app/database/dao_string.dart';
 
 class ProductPreferences extends ProductPreferencesManager with ChangeNotifier {
   ProductPreferences(
-    final ProductPreferencesSelection productPreferencesSelection,
-  ) : super(productPreferencesSelection);
+    final ProductPreferencesSelection productPreferencesSelection, {
+    this.daoString,
+  }) : super(productPreferencesSelection);
 
-  /// 2-letter language code of the latest reference load.
-  String _languageCode;
+  final DaoString? daoString;
 
-  /// "was it a network load" bool of the latest reference load.
-  bool _isNetwork;
+  /// Where we keep the language of the latest successful download.
+  static const String _DAO_STRING_KEY_LANGUAGE = 'latest_language';
 
-  String get languageCode => _languageCode;
-  bool get isNetwork => _isNetwork;
+  /// Were those data freshly downloaded? (and therefore good enough)
+  bool _isNetwork = false;
 
+  /// Is currently trying to download data?
+  bool _isDownloading = false;
+
+  /// Notifies listeners.
+  ///
+  /// Comments added only in order to avoid a "warning"
+  /// For the record, we need to override the method
+  /// because the parent's is protected
   @override
   void notifyListeners() => super.notifyListeners();
 
@@ -33,81 +43,178 @@ class ProductPreferences extends ProductPreferencesManager with ChangeNotifier {
   static String _getAttributeAssetPath(final String languageCode) =>
       'assets/metadata/init_attribute_groups_$languageCode.json';
 
-  /// Loads the references of importance and attribute groups from assets.
+  static const List<String> _DEFAULT_ATTRIBUTES = <String>[
+    Attribute.ATTRIBUTE_NUTRISCORE,
+    Attribute.ATTRIBUTE_ECOSCORE,
+    Attribute.ATTRIBUTE_NOVA,
+    Attribute.ATTRIBUTE_VEGETARIAN,
+    Attribute.ATTRIBUTE_VEGAN,
+    Attribute.ATTRIBUTE_PALM_OIL_FREE,
+    Attribute.ATTRIBUTE_LOW_SALT,
+    Attribute.ATTRIBUTE_LOW_SUGARS,
+    Attribute.ATTRIBUTE_LOW_FAT,
+    Attribute.ATTRIBUTE_LOW_SATURATED_FAT,
+    Attribute.ATTRIBUTE_LABELS_ORGANIC,
+    Attribute.ATTRIBUTE_LABELS_FAIR_TRADE,
+    Attribute.ATTRIBUTE_FOREST_FOOTPRINT,
+  ];
+
+  /// Inits with the best available not-network references.
   ///
-  /// May throw an exception.
-  Future<void> loadReferenceFromAssets(
-    final AssetBundle assetBundle, {
-    final String languageCode = _DEFAULT_LANGUAGE_CODE,
-  }) async {
-    final String importanceAssetPath = _getImportanceAssetPath(languageCode);
-    final String attributeGroupAssetPath = _getAttributeAssetPath(languageCode);
-    final String preferenceImportancesString =
-        await assetBundle.loadString(importanceAssetPath);
-    final String attributeGroupString =
-        await assetBundle.loadString(attributeGroupAssetPath);
-    final AvailableProductPreferences myAvailableProductPreferences =
-        AvailableProductPreferences.loadFromJSONStrings(
-      preferenceImportancesString: preferenceImportancesString,
-      attributeGroupsString: attributeGroupString,
-    );
-    availableProductPreferences = myAvailableProductPreferences;
-    _isNetwork = false;
-    _languageCode = languageCode;
+  /// That means trying with assets and local database.
+  Future<void> init(final AssetBundle assetBundle) async {
+    // trying the local database with the latest download language...
+    if (daoString != null) {
+      final String? latestLanguage =
+          await daoString!.get(_DAO_STRING_KEY_LANGUAGE);
+      if (latestLanguage != null) {
+        final bool successful = await _loadFromDatabase(latestLanguage);
+        if (successful) {
+          return;
+        }
+      }
+    }
+    // fallback: assets in English
+    await _loadFromAssets(assetBundle);
+  }
+
+  /// Refreshes the references with network data.
+  Future<void> refresh(final String languageCode) async {
+    if (_isNetwork) {
+      return;
+    }
+    if (_isDownloading) {
+      return;
+    }
+    _isDownloading = true;
+    final bool successful = await _loadFromNetwork(languageCode);
+    if (successful) {
+      _isNetwork = true;
+      if (daoString != null) {
+        await daoString!.put(_DAO_STRING_KEY_LANGUAGE, languageCode);
+      }
+      notify();
+    }
+    _isDownloading = false;
+  }
+
+  /// Loads the references of importance and attribute groups from assets.
+  Future<bool> _loadFromAssets(final AssetBundle assetBundle) async {
+    try {
+      const String languageCode = _DEFAULT_LANGUAGE_CODE;
+      final String importanceAssetPath = _getImportanceAssetPath(languageCode);
+      final String attributeGroupAssetPath =
+          _getAttributeAssetPath(languageCode);
+      final String preferenceImportancesString =
+          await assetBundle.loadString(importanceAssetPath);
+      final String attributeGroupsString =
+          await assetBundle.loadString(attributeGroupAssetPath);
+      _loadFromStrings(
+        languageCode,
+        preferenceImportancesString,
+        attributeGroupsString,
+      );
+      return true;
+    } catch (e) {
+      return false;
+    }
   }
 
   /// Loads the references of importance and attribute groups from urls.
   ///
+  /// The downloaded strings are automatically stored in the database.
+  Future<bool> _loadFromNetwork(String languageCode) async {
+    try {
+      final String importanceUrl =
+          AvailablePreferenceImportances.getUrl(languageCode);
+      final String attributeGroupUrl =
+          AvailableAttributeGroups.getUrl(languageCode);
+      final DownloadableString downloadableImportance;
+      downloadableImportance =
+          DownloadableString(Uri.parse(importanceUrl), dao: daoString);
+      final bool differentImportance = await downloadableImportance.download();
+      final DownloadableString downloadableAttributes =
+          DownloadableString(Uri.parse(attributeGroupUrl), dao: daoString);
+      final bool differentAttributes = await downloadableAttributes.download();
+      // the downloaded values are identical to what was stored locally.
+      if ((!differentImportance) && (!differentAttributes)) {
+        return false;
+      }
+      final String preferenceImportancesString = downloadableImportance.value!;
+      final String attributeGroupsString = downloadableAttributes.value!;
+      _loadFromStrings(
+        languageCode,
+        preferenceImportancesString,
+        attributeGroupsString,
+      );
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+
+  /// Loads the references of importance and attribute groups from database.
+  Future<bool> _loadFromDatabase(final String languageCode) async {
+    if (daoString == null) {
+      return false;
+    }
+    try {
+      final String importanceUrl =
+          AvailablePreferenceImportances.getUrl(languageCode);
+      final String attributeGroupUrl =
+          AvailableAttributeGroups.getUrl(languageCode);
+      final String? preferenceImportancesString =
+          await daoString!.get(importanceUrl);
+      final String? attributeGroupsString =
+          await daoString!.get(attributeGroupUrl);
+      if (preferenceImportancesString == null &&
+          attributeGroupsString == null) {
+        return false;
+      }
+      _loadFromStrings(
+        languageCode,
+        preferenceImportancesString!,
+        attributeGroupsString!,
+      );
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+
+  /// Loads the references of importance and attribute groups from strings.
+  ///
   /// May throw an exception.
-  Future<void> loadReferenceFromNetwork(String languageCode) async {
-    final String importanceUrl =
-        AvailablePreferenceImportances.getUrl(languageCode);
-    final String attributeGroupUrl =
-        AvailableAttributeGroups.getUrl(languageCode);
-    http.Response response;
-    response = await http.get(Uri.parse(importanceUrl));
-    if (response.statusCode != 200) {
-      return false;
-    }
-    final String preferenceImportancesString = response.body;
-    response = await http.get(Uri.parse(attributeGroupUrl));
-    if (response.statusCode != 200) {
-      return false;
-    }
-    final String attributeGroupsString = response.body;
+  void _loadFromStrings(
+    final String languageCode,
+    final String preferenceImportancesString,
+    final String attributeGroupsString,
+  ) {
     final AvailableProductPreferences myAvailableProductPreferences =
         AvailableProductPreferences.loadFromJSONStrings(
       preferenceImportancesString: preferenceImportancesString,
       attributeGroupsString: attributeGroupsString,
     );
     availableProductPreferences = myAvailableProductPreferences;
-    _isNetwork = true;
-    _languageCode = languageCode;
   }
 
   Future<void> resetImportances() async {
     await clearImportances(notifyListeners: false);
-    await setImportance(
-      Attribute.ATTRIBUTE_NUTRISCORE,
-      PreferenceImportance.ID_VERY_IMPORTANT,
-      notifyListeners: false,
-    );
-    await setImportance(
-      Attribute.ATTRIBUTE_NOVA,
-      PreferenceImportance.ID_IMPORTANT,
-      notifyListeners: false,
-    );
-    await setImportance(
-      Attribute.ATTRIBUTE_ECOSCORE,
-      PreferenceImportance.ID_IMPORTANT,
-      notifyListeners: false,
+    await Future.wait(
+      _DEFAULT_ATTRIBUTES.map(
+        (String attributeId) => setImportance(
+          attributeId,
+          PreferenceImportance.ID_IMPORTANT,
+          notifyListeners: false,
+        ),
+      ),
     );
     notify();
   }
 
   AttributeGroup getAttributeGroup(final String attributeId) {
-    for (final AttributeGroup attributeGroup in attributeGroups) {
-      for (final Attribute item in attributeGroup.attributes) {
+    for (final AttributeGroup attributeGroup in attributeGroups!) {
+      for (final Attribute item in attributeGroup.attributes!) {
         if (item.id == attributeId) {
           return attributeGroup;
         }
