@@ -1,14 +1,9 @@
-// Flutter imports:
 import 'package:flutter/material.dart';
-
-// Package imports:
 import 'package:openfoodfacts/model/Product.dart';
-
-// Project imports:
+import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/data_models/product_list_supplier.dart';
-import 'package:smooth_app/database/dao_product_list.dart';
-import 'package:smooth_app/database/local_database.dart';
+import 'package:smooth_app/database/product_query.dart';
 
 enum LoadingStatus {
   LOADING,
@@ -26,18 +21,24 @@ class ProductQueryModel with ChangeNotifier {
   final ProductListSupplier supplier;
 
   static const String _CATEGORY_ALL = 'all';
+  String currentCategory = _CATEGORY_ALL;
 
   LoadingStatus _loadingStatus = LoadingStatus.LOADING;
-  String _loadingError;
-  List<Product> _products;
-  List<Product> displayProducts;
-  bool isNotEmpty() => _products != null && _products.isNotEmpty;
+  String? _loadingError;
+  List<Product>? _products;
+  List<Product>? displayProducts;
+  bool isNotEmpty() => _products != null && _products!.isNotEmpty;
 
-  Map<String, String> categories = <String, String>{};
-  Map<String, int> categoriesCounter = <String, int>{};
-  List<String> sortedCategories;
+  /// <Label, Label (count)> [Map]
+  final Map<String, String> categories = <String, String>{};
 
-  String get loadingError => _loadingError;
+  /// <Label, count> [Map]
+  final Map<String, int> _categoriesCounter = <String, int>{};
+
+  /// Sorted labels
+  final List<String> sortedCategories = <String>[];
+
+  String? get loadingError => _loadingError;
   LoadingStatus get loadingStatus => _loadingStatus;
 
   Future<void> _asyncLoad() async {
@@ -51,8 +52,10 @@ class ProductQueryModel with ChangeNotifier {
     notifyListeners();
   }
 
-  /// Processes the data: stores it in database if needed, and lists categories
-  void process(final LocalDatabase localDatabase) {
+  /// Sorts the products by category.
+  ///
+  /// [translationForAll] is the displayed translation for meta category "All".
+  void process(final String translationForAll) {
     if (_loadingStatus != LoadingStatus.LOADED) {
       return;
     }
@@ -60,24 +63,21 @@ class ProductQueryModel with ChangeNotifier {
 
     final ProductList productList = supplier.getProductList();
     _products = productList.getList();
-    if (supplier.needsToBeSavedIntoDb()) {
-      DaoProductList(localDatabase).put(productList);
-    }
 
     displayProducts = _products;
 
-    categories[_CATEGORY_ALL] =
-        'All'; // TODO(monsieurtanuki): find a translation
+    categories[_CATEGORY_ALL] = translationForAll;
 
-    for (final Product product in _products) {
-      if (product.categoriesTags != null) {
-        for (final String category in product.categoriesTags) {
-          categories.putIfAbsent(category, () {
-            String title = category.substring(3).replaceAll('-', ' ');
-            title = '${title[0].toUpperCase()}${title.substring(1)}';
-            return title;
-          });
-          categoriesCounter[category] = (categoriesCounter[category] ?? 0) + 1;
+    for (final Product product in _products!) {
+      if (product.categoriesTagsInLanguages != null) {
+        final List<String>? translatedCategories =
+            product.categoriesTagsInLanguages![ProductQuery.getLanguage()];
+        if (translatedCategories != null) {
+          for (final String category in translatedCategories) {
+            categories[category] = '';
+            _categoriesCounter[category] =
+                (_categoriesCounter[category] ?? 0) + 1;
+          }
         }
       }
     }
@@ -86,34 +86,38 @@ class ProductQueryModel with ChangeNotifier {
 
     for (final String category in tempCategories) {
       if (category != _CATEGORY_ALL) {
-        if (categoriesCounter[category] <= 1) {
+        if (_categoriesCounter[category]! <= 1) {
           categories.remove(category);
         } else {
-          categories[category] =
-              '${categories[category]} (${categoriesCounter[category]})';
+          categories[category] = '$category (${_categoriesCounter[category]})';
         }
       }
     }
 
-    sortedCategories = categories.keys.toList();
+    sortedCategories.clear();
+    sortedCategories.addAll(categories.keys);
     sortedCategories.sort((String a, String b) {
       if (a == _CATEGORY_ALL) {
         return -1;
       } else if (b == _CATEGORY_ALL) {
         return 1;
       }
-      return categoriesCounter[b].compareTo(categoriesCounter[a]);
+      return _categoriesCounter[b]!.compareTo(_categoriesCounter[a]!);
     });
 
     _loadingStatus = LoadingStatus.COMPLETE;
   }
 
   void selectCategory(String category) {
+    currentCategory = category;
     if (category == _CATEGORY_ALL) {
       displayProducts = _products;
     } else {
-      displayProducts = _products
-          .where((Product product) => product.categoriesTags.contains(category))
+      displayProducts = _products!
+          .where((Product product) =>
+              product.categoriesTagsInLanguages?[ProductQuery.getLanguage()]
+                  ?.contains(category) ??
+              false)
           .toList();
     }
   }
