@@ -1,77 +1,108 @@
-// Flutter imports:
 import 'package:flutter/material.dart';
-
-// Package imports:
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:provider/provider.dart';
-import 'package:smooth_ui_library/widgets/smooth_toggle.dart';
-
-// Project imports:
 import 'package:smooth_app/data_models/continuous_scan_model.dart';
+import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/database/product_query.dart';
 import 'package:smooth_app/pages/scan/continuous_scan_page.dart';
+import 'package:smooth_app/pages/scan/ml_kit_scan_page.dart';
+import 'package:smooth_app/pages/scan/scanner_overlay.dart';
+import 'package:smooth_app/pages/user_preferences_dev_mode.dart';
 
-class ScanPage extends StatelessWidget {
-  const ScanPage({
-    @required this.contributionMode,
-  });
+class ScanPage extends StatefulWidget {
+  const ScanPage();
 
-  final bool contributionMode;
+  @override
+  State<ScanPage> createState() => _ScanPageState();
+}
+
+class _ScanPageState extends State<ScanPage> {
+  ContinuousScanModel? _model;
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _updateModel();
+  }
+
+  Future<void> _updateModel() async {
+    final LocalDatabase localDatabase = context.watch<LocalDatabase>();
+    if (_model == null) {
+      _model = await ContinuousScanModel().load(localDatabase);
+    } else {
+      await _model?.refresh();
+    }
+    setState(() {});
+  }
+
+  Future<PermissionStatus> _permissionCheck(
+    UserPreferences userPreferences,
+  ) async {
+    final PermissionStatus status = await Permission.camera.status;
+
+    // If is denied, is not restricted by for example parental control and is
+    // not already declined once
+    if (status.isDenied &&
+        !status.isRestricted &&
+        !userPreferences.cameraDeclinedOnce) {
+      final PermissionStatus newStatus = await Permission.camera.request();
+      if (!newStatus.isGranted && !newStatus.isLimited) {
+        userPreferences.setCameraDecline(true);
+      }
+      return newStatus;
+    } else {
+      return status;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
-    final LocalDatabase localDatabase = context.watch<LocalDatabase>();
-    return FutureBuilder<ContinuousScanModel>(
-        future: ContinuousScanModel(
-          contributionMode: contributionMode,
-          languageCode: ProductQuery.getCurrentLanguageCode(context),
-          countryCode: ProductQuery.getCurrentCountryCode(),
-        ).load(localDatabase),
-        builder: (BuildContext context,
-            AsyncSnapshot<ContinuousScanModel> snapshot) {
-          if (snapshot.connectionState == ConnectionState.done) {
-            final ContinuousScanModel continuousScanModel = snapshot.data;
-            if (continuousScanModel != null) {
-              return ContinuousScanPage(continuousScanModel);
-            }
-          }
+    final UserPreferences userPreferences = context.read<UserPreferences>();
+    if (_model == null) {
+      return const Center(child: CircularProgressIndicator());
+    }
+
+    return FutureBuilder<PermissionStatus>(
+      future: _permissionCheck(userPreferences),
+      builder: (
+        BuildContext context,
+        AsyncSnapshot<PermissionStatus> snapshot,
+      ) {
+        if (!snapshot.hasData) {
           return const Center(child: CircularProgressIndicator());
-        });
-  }
+        } else if (snapshot.hasError) {
+          return const Center(child: Text('Error'));
+        }
 
-  static Widget getContributeChooseToggle(
-          final ContinuousScanModel model, BuildContext context) =>
-      SmoothToggle(
-        value: model.contributionMode,
-        textLeft: '${AppLocalizations.of(context).scan_contribute}   ',
-        textRight: '     ${AppLocalizations.of(context).scan_choose}',
-        colorLeft: Colors.black.withAlpha(160),
-        colorRight: Colors.black.withAlpha(160),
-        iconLeft: SvgPicture.asset('assets/ikonate_bold/add.svg'),
-        iconRight: SvgPicture.asset('assets/ikonate_bold/search.svg'),
-        textSize: 12.0,
-        animationDuration: const Duration(milliseconds: 320),
-        width: 150.0,
-        height: 50.0,
-        onChanged: (bool value) => model.contributionModeSwitch(value),
-      );
+        // TODO(M123): show no camera access screen
+        if (snapshot.data!.isDenied ||
+            snapshot.data!.isPermanentlyDenied ||
+            snapshot.data!.isRestricted) {
+          const Center(
+            child: Text('No camera access granted'),
+          );
+        }
 
-  static Widget getHero(final Size screenSize) => Hero(
-        tag: 'action_button',
-        child: Container(
-          width: screenSize.width,
-          height: screenSize.height,
-          color: Colors.black,
-          child: Center(
-            child: SvgPicture.asset(
-              'assets/actions/scanner_alt_2.svg',
-              width: 60.0,
-              height: 60.0,
-              color: Colors.white,
+        final Widget child;
+
+        if (userPreferences.getFlag(
+              UserPreferencesDevMode.userPreferencesFlagUseMLKit,
+            ) ??
+            true) {
+          child = const MLKitScannerPage();
+        } else {
+          child = const ContinuousScanPage();
+        }
+
+        return ChangeNotifierProvider<ContinuousScanModel>(
+          create: (BuildContext context) => _model!,
+          child: Scaffold(
+            body: ScannerOverlay(
+              child: child,
             ),
           ),
-        ),
-      );
+        );
+      },
+    );
+  }
 }
