@@ -1,4 +1,3 @@
-// Project imports:
 import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/data_models/product_list_supplier.dart';
 import 'package:smooth_app/data_models/query_product_list_supplier.dart';
@@ -6,25 +5,19 @@ import 'package:smooth_app/database/dao_product_list.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/product_query.dart';
 
-class DatabaseProductListSupplier implements ProductListSupplier {
-  DatabaseProductListSupplier(this.productQuery, this.localDatabase);
-
-  final ProductQuery productQuery;
-  final LocalDatabase localDatabase;
-  ProductList _productList;
-
-  @override
-  ProductList getProductList() => _productList;
+class DatabaseProductListSupplier extends ProductListSupplier {
+  DatabaseProductListSupplier(
+    final ProductQuery productQuery,
+    final LocalDatabase localDatabase,
+    final int timestamp,
+  ) : super(productQuery, localDatabase, timestamp: timestamp);
 
   @override
-  Future<String> asyncLoad() async {
+  Future<String?> asyncLoad() async {
     try {
-      final ProductList productList = productQuery.getProductList();
-      final bool result = await DaoProductList(localDatabase).get(productList);
-      if (!result) {
-        return 'unexpected empty record';
-      }
-      _productList = productList;
+      final ProductList loadedProductList = productQuery.getProductList();
+      await DaoProductList(localDatabase).get(loadedProductList);
+      productList = loadedProductList;
       return null;
     } catch (e) {
       return e.toString();
@@ -32,9 +25,6 @@ class DatabaseProductListSupplier implements ProductListSupplier {
   }
 
   @override
-  bool needsToBeSavedIntoDb() => false;
-
-  @override
   ProductListSupplier getRefreshSupplier() =>
-      QueryProductListSupplier(productQuery);
+      QueryProductListSupplier(productQuery, localDatabase);
 }
