@@ -1,26 +1,27 @@
-// Package imports:
 import 'package:openfoodfacts/model/SearchResult.dart';
-
-// Project imports:
-import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/data_models/product_list_supplier.dart';
+import 'package:smooth_app/database/dao_product.dart';
+import 'package:smooth_app/database/dao_product_list.dart';
+import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/product_query.dart';
 
-class QueryProductListSupplier implements ProductListSupplier {
-  QueryProductListSupplier(this.productQuery);
-
-  final ProductQuery productQuery;
-  ProductList _productList;
-
-  @override
-  ProductList getProductList() => _productList;
+/// [ProductListSupplier] with a server query flavor
+class QueryProductListSupplier extends ProductListSupplier {
+  QueryProductListSupplier(
+    final ProductQuery productQuery,
+    final LocalDatabase localDatabase,
+  ) : super(productQuery, localDatabase);
 
   @override
-  Future<String> asyncLoad() async {
+  Future<String?> asyncLoad() async {
     try {
       final SearchResult searchResult = await productQuery.getSearchResult();
-      _productList = productQuery.getProductList();
-      _productList.addAll(searchResult.products);
+      productList = productQuery.getProductList();
+      if (searchResult.products != null) {
+        productList.setAll(searchResult.products!);
+        await DaoProduct(localDatabase).putAll(searchResult.products!);
+      }
+      await DaoProductList(localDatabase).put(productList);
       return null;
     } catch (e) {
       return e.toString();
@@ -28,8 +29,5 @@ class QueryProductListSupplier implements ProductListSupplier {
   }
 
   @override
-  bool needsToBeSavedIntoDb() => true;
-
-  @override
-  ProductListSupplier getRefreshSupplier() => null;
+  ProductListSupplier? getRefreshSupplier() => null;
 }
