@@ -1,58 +1,52 @@
-// Package imports:
 import 'package:openfoodfacts/model/Product.dart';
-
-// Project imports:
-import 'package:smooth_app/data_models/product_list.dart';
-import 'package:openfoodfacts/personalized_search/matched_product.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
+import 'package:smooth_app/data_models/user_preferences.dart';
+import 'package:smooth_app/helpers/smooth_matched_product.dart';
+
+/// Tabs where ranked products are displayed
+enum MatchTab {
+  YES,
+  MAYBE,
+  NO,
+  ALL,
+}
 
 class SmoothItModel {
-  static const int MATCH_INDEX_YES = 0;
-  static const int MATCH_INDEX_MAYBE = 1;
-  static const int MATCH_INDEX_NO = 2;
-  static const int MATCH_INDEX_ALL = 3;
-
-  final Map<int, List<MatchedProduct>> _categorizedProducts =
-      <int, List<MatchedProduct>>{};
-  List<MatchedProduct> _allProducts;
-  bool _nextRefreshIsJustChangingTabs = false;
+  final Map<MatchTab, List<MatchedProduct>> _categorizedProducts =
+      <MatchTab, List<MatchedProduct>>{};
 
   void refresh(
-    final ProductList productList,
+    final List<Product> products,
     final ProductPreferences productPreferences,
+    final UserPreferences userPreferences,
   ) {
-    if (_nextRefreshIsJustChangingTabs) {
-      _nextRefreshIsJustChangingTabs = false;
-      return;
-    }
-    final List<Product> unprocessedProducts = productList.getUniqueList();
-    _allProducts = MatchedProduct.sort(unprocessedProducts, productPreferences);
+    final List<MatchedProduct> allProducts = MatchedProduct.sort(
+      products,
+      productPreferences,
+      userPreferences,
+    );
     _categorizedProducts.clear();
-    for (final MatchedProduct matchededProduct in _allProducts) {
-      final int index = getMatchIndex(matchededProduct);
-      if (_categorizedProducts[index] == null) {
-        _categorizedProducts[index] = <MatchedProduct>[];
+    _categorizedProducts[MatchTab.ALL] = allProducts;
+    for (final MatchedProduct matchedProduct in allProducts) {
+      final MatchTab matchTab = _getMatchTab(matchedProduct);
+      if (_categorizedProducts[matchTab] == null) {
+        _categorizedProducts[matchTab] = <MatchedProduct>[];
       }
-      _categorizedProducts[index].add(matchededProduct);
+      _categorizedProducts[matchTab]!.add(matchedProduct);
     }
   }
 
-  void setNextRefreshAsJustChangingTabs() =>
-      _nextRefreshIsJustChangingTabs = true;
+  List<MatchedProduct> getMatchedProducts(final MatchTab matchTab) =>
+      _categorizedProducts[matchTab] ?? <MatchedProduct>[];
 
-  List<MatchedProduct> getMatchedProducts(final int matchIndex) =>
-      matchIndex == MATCH_INDEX_ALL
-          ? _allProducts
-          : _categorizedProducts[matchIndex] ?? <MatchedProduct>[];
-
-  static int getMatchIndex(final MatchedProduct matchedProduct) {
+  static MatchTab _getMatchTab(final MatchedProduct matchedProduct) {
     if ((matchedProduct.status == null) ||
         (matchedProduct.status == MatchedProductStatus.UNKNOWN)) {
-      return MATCH_INDEX_MAYBE;
-    } else if (matchedProduct.status == MatchedProductStatus.YES) {
-      return MATCH_INDEX_YES;
-    } else {
-      return MATCH_INDEX_NO;
+      return MatchTab.MAYBE;
     }
+    if (matchedProduct.status == MatchedProductStatus.YES) {
+      return MatchTab.YES;
+    }
+    return MatchTab.NO;
   }
 }
