@@ -1,165 +1,246 @@
-// Flutter imports:
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-
-// Package imports:
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:provider/provider.dart';
-
-// Project imports:
-import 'package:smooth_app/bottom_sheet_views/user_preferences_view.dart';
 import 'package:smooth_app/cards/product_cards/smooth_product_card_found.dart';
 import 'package:smooth_app/data_models/product_list.dart';
-import 'package:smooth_app/data_models/smooth_it_model.dart';
-import 'package:smooth_app/pages/product/common/product_query_page_helper.dart';
-import 'package:openfoodfacts/personalized_search/matched_product.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
-import 'package:smooth_app/themes/smooth_theme.dart';
+import 'package:smooth_app/data_models/smooth_it_model.dart';
+import 'package:smooth_app/data_models/user_preferences.dart';
+import 'package:smooth_app/database/dao_product_list.dart';
+import 'package:smooth_app/database/local_database.dart';
+import 'package:smooth_app/helpers/analytics_helper.dart';
+import 'package:smooth_app/helpers/smooth_matched_product.dart';
 
 class PersonalizedRankingPage extends StatefulWidget {
-  const PersonalizedRankingPage(this.productList);
+  PersonalizedRankingPage({
+    required final ProductList productList,
+    required this.title,
+  }) : products = productList.getList();
 
-  final ProductList productList;
+  const PersonalizedRankingPage.fromItems({
+    required this.products,
+    required this.title,
+  });
+
+  final List<Product> products;
+  final String title;
 
   @override
-  _PersonalizedRankingPageState createState() =>
+  State<PersonalizedRankingPage> createState() =>
       _PersonalizedRankingPageState();
 }
 
 class _PersonalizedRankingPageState extends State<PersonalizedRankingPage> {
-  static const List<int> _ORDERED_MATCH_INDEXES = <int>[
-    SmoothItModel.MATCH_INDEX_ALL,
-    SmoothItModel.MATCH_INDEX_YES,
-    SmoothItModel.MATCH_INDEX_MAYBE,
-    SmoothItModel.MATCH_INDEX_NO,
+  static const Map<MatchTab, Color> _COLORS = <MatchTab, Color>{
+    MatchTab.YES: Colors.green,
+    MatchTab.MAYBE: Colors.grey,
+    MatchTab.NO: Colors.red,
+    MatchTab.ALL: Colors.black,
+  };
+
+  static const List<MatchTab> _ORDERED_MATCH_TABS = <MatchTab>[
+    MatchTab.ALL,
+    MatchTab.YES,
+    MatchTab.NO,
+    MatchTab.MAYBE,
   ];
 
-  static const Map<int, IconData> _ICONS = <int, IconData>{
-    SmoothItModel.MATCH_INDEX_ALL: Icons.sort,
-    SmoothItModel.MATCH_INDEX_YES: Icons.check_circle,
-    SmoothItModel.MATCH_INDEX_MAYBE: CupertinoIcons.question_diamond,
-    SmoothItModel.MATCH_INDEX_NO: Icons.cancel,
-  };
-
-  static const Map<int, MaterialColor> _COLORS = <int, MaterialColor>{
-    SmoothItModel.MATCH_INDEX_YES: Colors.green,
-    SmoothItModel.MATCH_INDEX_MAYBE: Colors.grey,
-    SmoothItModel.MATCH_INDEX_NO: Colors.red,
-  };
-
   final SmoothItModel _model = SmoothItModel();
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
-
-  int _currentTabIndex = 0;
 
   @override
   Widget build(BuildContext context) {
     final ProductPreferences productPreferences =
         context.watch<ProductPreferences>();
-    final ColorScheme colorScheme = Theme.of(context).colorScheme;
-    _model.refresh(widget.productList, productPreferences);
-    final List<BottomNavigationBarItem> bottomNavigationBarItems =
-        <BottomNavigationBarItem>[];
-    final AppLocalizations appLocalizations = AppLocalizations.of(context);
-    for (final int matchIndex in _ORDERED_MATCH_INDEXES) {
-      bottomNavigationBarItems.add(
-        BottomNavigationBarItem(
-          icon: Icon(
-            _ICONS[matchIndex],
-            color: _COLORS[matchIndex] == null
-                ? null
-                : SmoothTheme.getColor(
-                    colorScheme,
-                    _COLORS[matchIndex],
-                    ColorDestination.SURFACE_FOREGROUND,
-                  ),
-          ),
-          label: _model.getMatchedProducts(matchIndex).length.toString(),
-        ),
+    final LocalDatabase localDatabase = context.watch<LocalDatabase>();
+    final DaoProductList daoProductList = DaoProductList(localDatabase);
+    _model.refresh(
+      widget.products,
+      productPreferences,
+      context.watch<UserPreferences>(),
+    );
+    final AppLocalizations appLocalizations = AppLocalizations.of(context)!;
+    final List<Color> colors = <Color>[];
+    final List<String> titles = <String>[];
+    final List<List<MatchedProduct>> matchedProductsList =
+        <List<MatchedProduct>>[];
+    for (final MatchTab matchTab in _ORDERED_MATCH_TABS) {
+      final List<MatchedProduct> products = _model.getMatchedProducts(matchTab);
+      matchedProductsList.add(products);
+      titles.add(
+        matchTab == MatchTab.ALL
+            ? appLocalizations.ranking_tab_all
+            : products.length.toString(),
       );
+      colors.add(_COLORS[matchTab]!);
     }
-    return Scaffold(
-      key: _scaffoldKey,
-      appBar: AppBar(
-        title: Row(
-          mainAxisSize: MainAxisSize.max,
-          children: <Widget>[
-            Flexible(
-              child: Text(
-                ProductQueryPageHelper.getProductListLabel(
-                  widget.productList,
-                  context,
-                ),
-                overflow: TextOverflow.fade,
+
+    AnalyticsHelper.trackPersonalizedRanking(
+      title: widget.title,
+      products: matchedProductsList[0].length,
+      goodProducts: matchedProductsList[1].length,
+      badProducts: matchedProductsList[2].length,
+      unknownProducts: matchedProductsList[3].length,
+    );
+
+    return DefaultTabController(
+      length: _ORDERED_MATCH_TABS.length,
+      initialIndex: _ORDERED_MATCH_TABS.indexOf(MatchTab.YES),
+      child: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.white,
+          foregroundColor: Colors.black,
+          bottom: TabBar(
+            unselectedLabelStyle: const TextStyle(
+              fontSize: 15,
+            ),
+            labelStyle: const TextStyle(
+              fontSize: 20,
+              decoration: TextDecoration.underline,
+            ),
+            indicator: const BoxDecoration(
+              border: Border(
+                left: BorderSide(color: Colors.grey), // provides to left side
+                right: BorderSide(color: Colors.grey), // for right side
               ),
             ),
-          ],
-        ),
-        actions: <Widget>[
-          IconButton(
-            icon: const Icon(Icons.settings),
-            onPressed: () => UserPreferencesView.showModal(
-              context,
-              callback: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(
-                    content:
-                        Text(appLocalizations.reloaded_with_new_preferences),
-                    duration: const Duration(milliseconds: 1500),
+            isScrollable: false,
+            tabs: <Tab>[
+              ...List<Tab>.generate(
+                _ORDERED_MATCH_TABS.length,
+                (final int index) => Tab(
+                  child: Text(
+                    titles[index],
+                    style: TextStyle(color: colors[index]),
                   ),
-                );
-              },
+                ),
+              ),
+            ],
+          ),
+          title: Row(
+            mainAxisSize: MainAxisSize.max,
+            children: <Widget>[
+              Flexible(
+                child: Text(widget.title, overflow: TextOverflow.fade),
+              ),
+            ],
+          ),
+        ),
+        body: TabBarView(
+          children: List<Widget>.generate(
+            _ORDERED_MATCH_TABS.length,
+            (final int index) => _getStickyHeader(
+              _ORDERED_MATCH_TABS[index],
+              matchedProductsList[index],
+              appLocalizations,
+              daoProductList,
             ),
-          )
-        ],
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        type: BottomNavigationBarType.fixed,
-        currentIndex: _currentTabIndex,
-        items: bottomNavigationBarItems,
-        onTap: (int tapped) => setState(() {
-          _currentTabIndex = tapped;
-          _model.setNextRefreshAsJustChangingTabs();
-        }),
-      ),
-      body: _getStickyHeader(
-        _model.getMatchedProducts(_ORDERED_MATCH_INDEXES[_currentTabIndex]),
-        colorScheme,
-        appLocalizations,
+          ),
+        ),
       ),
     );
   }
 
   Widget _buildSmoothProductCard(
     final MatchedProduct matchedProduct,
-    final ColorScheme colorScheme,
+    final DaoProductList daoProductList,
+    final AppLocalizations appLocalizations,
   ) =>
-      Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 8.0),
-        child: SmoothProductCardFound(
-          heroTag: matchedProduct.product.barcode,
-          product: matchedProduct.product,
-          elevation: 4.0,
-          backgroundColor: SmoothTheme.getColor(
-            colorScheme,
-            _COLORS[SmoothItModel.getMatchIndex(matchedProduct)],
-            ColorDestination.SURFACE_BACKGROUND,
+      Dismissible(
+        key: Key(matchedProduct.product.barcode!),
+        onDismissed: (final DismissDirection direction) async {
+          final bool removed = widget.products.remove(matchedProduct.product);
+          if (removed) {
+            setState(() {});
+          }
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text(
+                removed
+                    ? appLocalizations.product_removed
+                    : appLocalizations.product_could_not_remove,
+              ),
+              duration: const Duration(seconds: 3),
+            ),
+          );
+        },
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 8.0),
+          child: SmoothProductCardFound(
+            heroTag: matchedProduct.product.barcode!,
+            product: matchedProduct.product,
+            elevation: 4.0,
           ),
         ),
       );
 
   Widget _getStickyHeader(
-          final List<MatchedProduct> matchedProducts,
-          final ColorScheme colorScheme,
-          final AppLocalizations appLocalizations) =>
-      matchedProducts.isEmpty
-          ? Center(
-              child: Text(appLocalizations.no_product_in_section,
-                  style: Theme.of(context).textTheme.subtitle1),
-            )
-          : ListView.builder(
-              itemCount: matchedProducts.length,
-              itemBuilder: (BuildContext context, int index) =>
-                  _buildSmoothProductCard(matchedProducts[index], colorScheme),
+    final MatchTab matchTab,
+    final List<MatchedProduct> matchedProducts,
+    final AppLocalizations appLocalizations,
+    final DaoProductList daoProductList,
+  ) {
+    final Widget? subtitleWidget = _getSubtitleWidget(
+      _COLORS[matchTab],
+      _getSubtitle(matchTab, appLocalizations),
+    );
+    if (matchedProducts.isEmpty) {
+      return Column(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: <Widget>[
+          subtitleWidget ?? Container(),
+          Text(
+            appLocalizations.no_product_in_section,
+            style: Theme.of(context).textTheme.subtitle1,
+          ),
+          Container(),
+        ],
+      );
+    }
+    final int additional = subtitleWidget == null ? 0 : 1;
+    return ListView.builder(
+      itemCount: matchedProducts.length + additional,
+      itemBuilder: (BuildContext context, int index) => index < additional
+          ? subtitleWidget!
+          : _buildSmoothProductCard(
+              matchedProducts[index - additional],
+              daoProductList,
+              appLocalizations,
+            ),
+    );
+  }
+
+  Widget? _getSubtitleWidget(
+    final Color? color,
+    final String? subtitle,
+  ) =>
+      subtitle == null
+          ? null
+          : Container(
+              padding: const EdgeInsets.all(8),
+              child: Center(
+                child: Text(
+                  subtitle,
+                  style: const TextStyle(color: Colors.white),
+                ),
+              ),
+              color: color,
             );
+
+  String? _getSubtitle(
+    final MatchTab matchTab,
+    final AppLocalizations appLocalizations,
+  ) {
+    switch (matchTab) {
+      case MatchTab.ALL:
+        return null;
+      case MatchTab.MAYBE:
+        return appLocalizations.ranking_subtitle_match_maybe;
+      case MatchTab.YES:
+        return appLocalizations.ranking_subtitle_match_yes;
+      case MatchTab.NO:
+        return appLocalizations.ranking_subtitle_match_no;
+    }
+  }
 }
