@@ -1,111 +1,66 @@
-// Flutter imports:
 import 'package:flutter/material.dart';
-
-// Package imports:
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:openfoodfacts/model/Product.dart';
-import 'package:smooth_ui_library/buttons/smooth_simple_button.dart';
-import 'package:smooth_ui_library/dialogs/smooth_alert_dialog.dart';
-
-// Project imports:
+import 'package:openfoodfacts/openfoodfacts.dart';
+import 'package:smooth_app/data_models/fetched_product.dart';
 import 'package:smooth_app/database/barcode_product_query.dart';
 import 'package:smooth_app/database/dao_product.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/database/product_query.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_action_button.dart';
+import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_app/generic_lib/loading_dialog.dart';
 
 /// Dialog helper for product barcode search
 class ProductDialogHelper {
   ProductDialogHelper({
-    this.barcode,
-    this.context,
-    this.localDatabase,
-    this.refresh,
+    required this.barcode,
+    required this.context,
+    required this.localDatabase,
+    required this.refresh,
   });
 
   final String barcode;
   final BuildContext context;
   final LocalDatabase localDatabase;
   final bool refresh;
-  bool _popEd = false;
 
-  Future<Product> openBestChoice() async {
-    final Product product = await DaoProduct(localDatabase).get(barcode);
+  Future<FetchedProduct> openBestChoice() async {
+    final Product? product = await DaoProduct(localDatabase).get(barcode);
     if (product != null) {
-      return product;
+      return FetchedProduct(product);
     }
     return openUniqueProductSearch();
   }
 
-  Future<Product> openUniqueProductSearch() => showDialog<Product>(
-        context: context,
-        builder: (BuildContext context) {
-          BarcodeProductQuery(
+  Future<FetchedProduct> openUniqueProductSearch() async =>
+      await LoadingDialog.run<FetchedProduct>(
+          context: context,
+          future: BarcodeProductQuery(
             barcode: barcode,
-            languageCode: ProductQuery.getCurrentLanguageCode(context),
-            countryCode: ProductQuery.getCurrentCountryCode(),
-          ).getProduct().then((Product value) async {
-            if (value != null) {
-              await DaoProduct(localDatabase).put(value);
-            }
-            _popSearchingDialog(value);
-            /* TODO(monsieurtanuki): better granularity - being able to say
-             1. you clicked on 'stop'
-             2. no internet connection
-             3. no result at all
-             4. time out
-             5. of course, the product (when everything is fine)
-             */
-          });
-          return _getSearchingDialog();
-        },
-      );
+            daoProduct: DaoProduct(localDatabase),
+          ).getFetchedProduct(),
+          title: refresh
+              ? AppLocalizations.of(context)!.refreshing_product
+              : '${AppLocalizations.of(context)!.looking_for}: $barcode') ??
+      FetchedProduct.error(FetchedProductStatus.userCancelled);
 
-  void _popSearchingDialog(final Product product) {
-    if (_popEd) {
-      return;
-    }
-    _popEd = true;
-    Navigator.pop(context, product);
-  }
-
-  Widget _getSearchingDialog() => SmoothAlertDialog(
-        close: false,
-        body: ListTile(
-          leading: const CircularProgressIndicator(),
-          title: Text(
-            refresh
-                ? AppLocalizations.of(context).refreshing_product
-                : '${AppLocalizations.of(context).looking_for}: $barcode',
-          ),
-        ),
-        actions: <SmoothSimpleButton>[
-          SmoothSimpleButton(
-            text: AppLocalizations.of(context).stop,
-            important: false,
-            onPressed: () => _popSearchingDialog(null),
-          ),
-        ],
-      );
-
-  void openProductNotFoundDialog() => showDialog<Widget>(
+  void _openProductNotFoundDialog() => showDialog<Widget>(
         context: context,
         builder: (BuildContext context) {
           return SmoothAlertDialog(
             close: false,
             body: Text(
               refresh
-                  ? AppLocalizations.of(context).could_not_refresh
-                  : '${AppLocalizations.of(context).no_product_found}: $barcode',
+                  ? AppLocalizations.of(context)!.could_not_refresh
+                  : '${AppLocalizations.of(context)!.no_product_found}: $barcode',
             ),
-            actions: <SmoothSimpleButton>[
-              SmoothSimpleButton(
-                text: AppLocalizations.of(context).close,
-                important: false,
+            actions: <SmoothActionButton>[
+              SmoothActionButton(
+                text: AppLocalizations.of(context)!.close,
                 onPressed: () => Navigator.pop(context),
               ),
-              SmoothSimpleButton(
-                text: AppLocalizations.of(context).contribute,
-                important: true,
+              SmoothActionButton(
+                text: AppLocalizations.of(context)!.contribute,
+
                 onPressed: () => Navigator.pop(
                     context), // TODO(monsieurtanuki): to be implemented
               ),
@@ -113,4 +68,41 @@ class ProductDialogHelper {
           );
         },
       );
+
+  static Widget getErrorMessage(final String message) => ListTile(
+        leading: const Icon(Icons.error_outline, color: Colors.red),
+        title: Text(message),
+      );
+
+  void _openErrorMessage(final String message) => showDialog<void>(
+        context: context,
+        builder: (BuildContext context) => SmoothAlertDialog(
+          close: false,
+          body: getErrorMessage(message),
+          actions: <SmoothActionButton>[
+            SmoothActionButton(
+              text: AppLocalizations.of(context)!.close,
+              onPressed: () => Navigator.pop(context),
+            ),
+          ],
+        ),
+      );
+
+  /// Opens an error dialog; to be used only if the status is not ok.
+  void openError(final FetchedProduct fetchedProduct) {
+    final AppLocalizations appLocalizations = AppLocalizations.of(context)!;
+    switch (fetchedProduct.status) {
+      case FetchedProductStatus.ok:
+        throw Exception("You're not supposed to call this if the status is ok");
+      case FetchedProductStatus.userCancelled:
+        _openErrorMessage(appLocalizations.product_internet_cancel);
+        return;
+      case FetchedProductStatus.internetError:
+        _openErrorMessage(appLocalizations.product_internet_error);
+        return;
+      case FetchedProductStatus.internetNotFound:
+        _openProductNotFoundDialog();
+        return;
+    }
+  }
 }
