@@ -1,16 +1,24 @@
-// Flutter imports:
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class ConstantIcons {
+  @visibleForTesting
+  const ConstantIcons();
+
+  static ConstantIcons get instance => _instance ??= const ConstantIcons();
+  static ConstantIcons? _instance;
+
+  /// Setter that allows tests to override the singleton instance.
+  @visibleForTesting
+  static set instance(ConstantIcons testInstance) => _instance = testInstance;
+
   static bool _isApple() =>
       defaultTargetPlatform == TargetPlatform.iOS ||
       defaultTargetPlatform == TargetPlatform.macOS;
 
-  static IconData getShareIcon() =>
+  IconData getShareIcon() =>
       _isApple() ? CupertinoIcons.square_arrow_up : Icons.share;
 
-  static IconData getBackIcon() =>
-      _isApple() ? CupertinoIcons.back : Icons.arrow_back;
+  IconData getBackIcon() => _isApple() ? CupertinoIcons.back : Icons.arrow_back;
 }
